requirements.txt
pandas
numpy<2.0.0
matplotlib
seaborn
plotly
scikit-learn
scipy
tensorflow
streamlit-ydata-profiling
kaggle